More Status Menu
- upped data type of MenuData::_currentNumber from byte to int
- put in some logic in the menu to loop through zones while displaying
status

+++ src/main/arduino_fuoco/entity/MenuData.cpp
@@ -47,12 +47,12 @@ namespace ArduinoFuoco
       return &_currentTime;
     }
 
-    void MenuData::setCurrentNumber(byte number)
+    void MenuData::setCurrentNumber(int number)
     {
       _currentNumber = number;
     }
 
-    byte MenuData::getCurrentNumber() const
+    int MenuData::getCurrentNumber() const
     {
       return _currentNumber;
     }
